Ticket #2412 Fix for incorrect tabs in KitchenSink
Fixed race condition between setActiveTab and open call for tabGroupProxy that caused the the initial tab to not be set correctly if the tabGroup had not finished opening.  If the tabGroup is still opening when setActiveTab is called, the initialActiveTab value will be set so that handlePostOpen will correctly pick up and utilize the value.  Should also fix #2071

Also changed the app.jss to include the titleid for the first tab so the label shows correctly.