Updated Ti.include description based on bugs TIMOB-4108 and TIMOB-2358.
Moved Global objects documentation from the notes in the Titanium module to Global.yml to reduce duplication. Need to revisit the Global.yml file later to organize the information properly, break out the Global.JSON module, and move the String method descriptions to the appropriate place. But let's not complicate things unnecessarily.