Updated to address comments on PR 676.
Added a few default values. Also, realized that textAlign was missing from TextArea and TextField, although it is used in KS, and seems to work on both iOS and Android.

In the course of checking button defaults, realized that the Button code sample is so simple, the button it creates isn't recognizable as a button.