When targetSdkVersion is set in tiapp.xml's <android><manifest> section, build apps against that version (or the next best previous version available in the user's Android SDK folder.)
If the resulting SDK version is Honeycomb MR2 (3.2, api level 13) or higher, add "screenSize" to the list of "configChanges" that our Activities handle.  This is necessary because when targeting api level 13 or higher, a change in device orientation causes a screenSize config change that is not automatically handled by Android and thus causes the activity to restart unless "configChanges" contains an entry for screenSize.

[TIMOB-10863]